+++ src/main/ruby/truffleruby/core/exception.rb
@@ -211,18 +211,19 @@ end
 
 class NameError < StandardError
 
-  def initialize(*args)
+  def initialize(*args, receiver: undefined)
     name = args.size > 1 ? args.pop : nil
     super(*args)
     TrufflePrimitive.name_error_set_name self, name
+    TrufflePrimitive.name_error_set_receiver self, receiver unless TrufflePrimitive.undefined?(receiver)
   end
 end
 
 class NoMethodError < NameError
 
-  def initialize(*arguments)
+  def initialize(*arguments, receiver: undefined)
     args = arguments.size > 2 ? arguments.pop : nil
-    super(*arguments) # TODO BJF Jul 24, 2016 Need to handle NoMethodError.new(1,2,3,4)
+    super(*arguments, receiver: receiver) # TODO BJF Jul 24, 2016 Need to handle NoMethodError.new(1,2,3,4)
     TrufflePrimitive.no_method_error_set_args self, args
   end
 end